@@ -42,7 +42,7 @@ public class ViewCardSetFragment extends Fragment {
         ArrayAdapter<String> autoCompleteAdapter=
                 new ArrayAdapter<String>(getContext(),android.R.layout.simple_dropdown_item_1line,
                         viewCardSetViewModel.getSetNamesDropdownList());
-        AutoCompleteTextView textView=(AutoCompleteTextView)root.findViewById(R.id.setSearch);
+        AutoCompleteTextView textView= root.findViewById(R.id.setSearch);
         textView.setThreshold(3);
         textView.setAdapter(autoCompleteAdapter);
 
@@ -60,7 +60,7 @@ public class ViewCardSetFragment extends Fragment {
 
         binding.setSearch.addTextChangedListener(new ViewCardSet_SetSearchBarChangedListener(binding.setSearch, viewCardSetViewModel, adapter, layout));
 
-        viewCardSetViewModel.getDbRefreshIndicator().observe(this, new Observer<Boolean>() {
+        viewCardSetViewModel.getDbRefreshIndicator().observe(getViewLifecycleOwner(), new Observer<Boolean>() {
             @Override
             public void onChanged(Boolean aBoolean) {
                 if(aBoolean){
