@@ -70,7 +70,7 @@ public class ViewCardsSummaryFragment extends Fragment {
             });
         }
 
-        viewCardsViewModel.getDbRefreshIndicator().observe(this, new Observer<Boolean>() {
+        viewCardsViewModel.getDbRefreshIndicator().observe(getViewLifecycleOwner(), new Observer<Boolean>() {
             @Override
             public void onChanged(Boolean aBoolean) {
                 if(aBoolean){
